+++ src/basic_pub_py.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python
 import rospy
-from std_msgs.msg import Int32
+from std_msgs.msg import Int64
 
 def main_pub():
     rospy.init_node('basic_publisher', anonymous=False)
-    basic_pub = rospy.Publisher('test_msg', Int32, queue_size=1)
-    msg = Int32(); count = 0
+    basic_pub = rospy.Publisher('test_msg', Int64, queue_size=1)
+    msg = Int64(); count = 0
     r = rospy.Rate(1)
 
     while not rospy.is_shutdown():
