+++ src/basic_sub_cpp.cpp
@@ -1,8 +1,8 @@
 #include "ros/ros.h"
-#include "std_msgs/Int32.h"
+#include "std_msgs/Int64.h"
 
-void basicCallback(const std_msgs::Int32::ConstPtr& msg) {
-    ROS_INFO("[BasicSubscriber] Receive data : %d", msg->data);
+void basicCallback(const std_msgs::Int64::ConstPtr& msg) {
+    ROS_INFO("[BasicSubscriber] Receive data : %ld", msg->data);
 }
 
 int main(int argc, char **argv) {
